requirements.txt
annotated-types==0.7.0
anyio==4.12.0
blinker==1.7.0
cassandra-driver==3.29.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.7
distro==1.9.0
Flask==3.0.0
geomet==1.1.0
git-filter-repo==2.47.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.1.2
Jinja2==3.1.2
jiter==0.12.0
MarkupSafe==2.1.3
mysql-connector-python==9.5.0
openai==2.14.0
packaging==25.0
pillow==12.0.0
pydantic==2.12.5
pydantic_core==2.41.5
RapidFuzz==3.14.3
requests==2.32.5
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
Werkzeug==3.0.1
